LPS-179567 Extract CommercePriceListEntriesScreenNavigationCategory methods to a CommercePriceListEntriesScreenNavigationEntry component and extends CommercePriceListEntriesScreenNavigationCategory in it
