+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/frontend/taglib/servlet/taglib/CommercePriceListEntriesScreenNavigationCategory.java
@@ -14,31 +14,11 @@
 
 package com.liferay.commerce.pricing.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.price.list.model.CommercePriceList;
-import com.liferay.commerce.price.list.service.CommercePriceEntryService;
-import com.liferay.commerce.price.list.service.CommercePriceListService;
 import com.liferay.commerce.pricing.web.internal.constants.CommercePriceListScreenNavigationConstants;
-import com.liferay.commerce.pricing.web.internal.display.context.CommercePriceEntryDisplayContext;
-import com.liferay.commerce.product.service.CommerceCatalogService;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -47,15 +27,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=30",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=30",
+	service = ScreenNavigationCategory.class
 )
 public class CommercePriceListEntriesScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommercePriceList> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -63,13 +39,8 @@ public class CommercePriceListEntriesScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return CommercePriceListScreenNavigationConstants.CATEGORY_KEY_ENTRIES;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(
+		return language.get(
 			locale,
 			CommercePriceListScreenNavigationConstants.CATEGORY_KEY_ENTRIES);
 	}
@@ -80,81 +51,7 @@ public class CommercePriceListEntriesScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_PRICE_LIST_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CommercePriceList commercePriceList) {
-		if (commercePriceList == null) {
-			return false;
-		}
-
-		boolean hasPermission = false;
-
-		try {
-			hasPermission = _commercePriceListModelResourcePermission.contains(
-				PermissionThreadLocal.getPermissionChecker(),
-				commercePriceList.getCommercePriceListId(), ActionKeys.VIEW);
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-		}
-
-		return hasPermission;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		try {
-			CommercePriceEntryDisplayContext commercePriceEntryDisplayContext =
-				new CommercePriceEntryDisplayContext(
-					_commerceCatalogService, _commercePriceEntryService,
-					_commercePriceListModelResourcePermission,
-					_commercePriceListService, httpServletRequest);
-
-			httpServletRequest.setAttribute(
-				WebKeys.PORTLET_DISPLAY_CONTEXT,
-				commercePriceEntryDisplayContext);
-		}
-		catch (Exception exception) {
-			_log.error(exception);
-		}
-
-		_jspRenderer.renderJSP(
-			_setServletContext, httpServletRequest, httpServletResponse,
-			"/commerce_price_lists/price_entries.jsp");
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CommercePriceListEntriesScreenNavigationCategory.class);
-
 	@Reference
-	private CommerceCatalogService _commerceCatalogService;
-
-	@Reference
-	private CommercePriceEntryService _commercePriceEntryService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.commerce.price.list.model.CommercePriceList)"
-	)
-	private ModelResourcePermission<CommercePriceList>
-		_commercePriceListModelResourcePermission;
-
-	@Reference
-	private CommercePriceListService _commercePriceListService;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.pricing.web)"
-	)
-	private ServletContext _setServletContext;
+	protected Language language;
 
 }
